Move useSearch into its own file
